fix: class renamed to ProcessOrderPaymentUseCase
